Final refaktor
