+++ MainMenu.rb
@@ -1,8 +1,10 @@
 load './Search.rb'
 require './Fridge.rb'
+require './recipe_act.rb'
+require './recipe_store.rb'
 
 class Menu_main
-  def self.general_main
+  def self.main
     while true
       puts '1. Recipe'
       puts '2. Fridge'
@@ -15,48 +17,21 @@ class Menu_main
       when '0'
         exit
       when '1'
-        menu_of_recipe
+        Recipe.menu
       when '2'
-        Fridge.menu_of_fridge
+        Fridge.menu
       when '3'
-        Search.menu_of_search
+        Search.menu
+      else
+        error_text
       end
     end
   end
-
-  def self.menu_of_recipe
-    while true
-      puts '1. Show all'
-      puts '2. Search by Name'
-      puts '3. Create recipe'
-      puts '4. Show Favorites'
-      puts '0. Back'
-
-      puts 'Enter a number '
-      user_inputs = gets.chomp
-      case user_inputs
-      when '1'
-        DATA.all_recipes.each { |recipe| puts "Recipe Name: #{recipe.name}, ID:#{recipe.id}" }
-        RecipeAct.id_from_user
-      when '2'
-        Search.by_name
-        RecipeAct.id_from_user
-      when '3'
-        Recipe.add_recipe
-      when '4'
-        DATA.all_recipes.each do |recipe|
-          if recipe.is_favorite
-            puts "Recipe Name: #{recipe.name}, ID:#{recipe.id}"
-          end
-        end
-        RecipeAct.id_from_user
-      when '0'
-        general_main
-      else
-        puts "Enter correct number"
-      end
-    end
+  def self.error_text
+    puts "*" * 24
+    puts format("| %12s | ",  "Enter correct number" )
+    puts "*" * 24
   end
 end
 
-Menu_main.general_main
+Menu_main.main
