+++ fridge.json
@@ -1,1 +1,1 @@
-{"milk":7.0,"apple":3.0,"1":2.0,"234":1.0,"1234":2.0,"banana ":2.0,"табір":1.0,"церква":1.0}
+{"Potato":0.0,"":0.0}
